fix(studio): Warns user he will be logged out (#185)
When the a 401 (unauthorized) is detected, a warning is shown to the user that they are about to be logged out, and then they are logged out. 

Current behaviour, is whatever action triggered the 401 will show an error message independently (unless it's caught and hidden), and nothing else will happen, until the minute timer detects the user has logged out.